feat: add BackgroundDoodles component for animated decorative images and integrate into Home and Learn pages
